move "terminal things" from kernel.c to terminal.c
