+++ workdir/kernel.c
@@ -1,97 +1,4 @@
-/* include type definition for size_t, utint, etc ... */
-#include <stddef.h> // size_t
-#include <stdint.h> // uint?_t
-#include <system.h>
-
-/* Hardware text mode color constants. */
-enum vga_color {
-  COLOR_BLACK = 0,
-  COLOR_BLUE = 1,
-  COLOR_GREEN = 2,
-  COLOR_CYAN = 3,
-  COLOR_RED = 4,
-  COLOR_MAGENTA = 5,
-  COLOR_BROWN = 6,
-  COLOR_LIGHT_GREY = 7,
-  COLOR_DARK_GREY = 8,
-  COLOR_LIGHT_BLUE = 9,
-  COLOR_LIGHT_GREEN = 10,
-  COLOR_LIGHT_CYAN = 11,
-  COLOR_LIGHT_RED = 12,
-  COLOR_LIGHT_MAGENTA = 13,
-  COLOR_LIGHT_BROWN = 14,
-  COLOR_WHITE = 15,
-};
-
-uint8_t make_color(enum vga_color fg, enum vga_color bg) {
-  return fg | bg << 4;
-}
-
-uint16_t make_vgaentry(char c, uint8_t color) {
-  uint16_t c16 = c;
-  uint16_t color16 = color;
-  return c16 | color16 << 8;
-}
-
-static const size_t VGA_WIDTH = 80;
-static const size_t VGA_HEIGHT = 25;
-
-size_t terminal_row;
-size_t terminal_column;
-
-/*
-  each entry in terminal is represented by 2byte
-    - ASCII character takes one byte
-    - color information takes other byte
-  see above `make_vgaentry`
-*/
-uint8_t terminal_color;
-uint16_t* terminal_buffer;
-
-void terminal_initialize() {
-  terminal_row = 0;
-  terminal_column = 0;
-  terminal_color = make_color(COLOR_LIGHT_GREY, COLOR_BLACK);
-
-  /* VGA text mode buffer is located at 0xB8000 */
-  terminal_buffer = (uint16_t*) 0xB8000;
-  for (size_t y = 0; y < VGA_HEIGHT; y++) {
-    for (size_t x = 0; x < VGA_WIDTH; x++) {
-      const size_t index = y * VGA_WIDTH + x;
-      terminal_buffer[index] = make_vgaentry(' ', terminal_color);
-    }
-  }
-}
-
-void terminal_setcolor(uint8_t color) {
-  terminal_color = color;
-}
-
-void terminal_putentryat(char c, uint8_t color, size_t x, size_t y) {
-  const size_t index = y * VGA_WIDTH + x;
-  terminal_buffer[index] = make_vgaentry(c, color);
-}
-
-void terminal_putchar(char c) {
-  if (c == '\n') {
-    terminal_column = 0;
-    terminal_row++;
-  } else {
-    terminal_putentryat(c, terminal_color, terminal_column, terminal_row);
-    if (++terminal_column == VGA_WIDTH) {
-      terminal_column = 0;
-      if (++terminal_row == VGA_HEIGHT) {
-        terminal_row = 0;
-      }
-    }
-  }
-}
-
-void terminal_writestring(const char* data) {
-  size_t datalen = strlen(data);
-  for (size_t i = 0; i < datalen; i++)
-    terminal_putchar(data[i]);
-}
+#include <system.h> // terminal_initialize, terminal_writestring
 
 void kernel_main() {
   terminal_initialize();
